+++ app/controllers/upright/public/incidents_controller.rb
@@ -1,6 +1,6 @@
 class Upright::Public::IncidentsController < Upright::Public::BaseController
   def show
-    @incident = Upright::Incident.find(params[:id])
+    @incident = Upright::Incident.public_facing.find(params[:id])
     expires_in 15.seconds, public: true
   end
 end