+++ test/integration/prometheus_proxy_controller_test.rb
@@ -11,7 +11,7 @@ class PrometheusProxyControllerTest < ActionDispatch::IntegrationTest
     stub_request(:get, "http://localhost:9090/graph").to_return(status: 200, body: "Prometheus UI")
     sign_in
 
-    get "/prometheus/graph"
+    get "/prometheus/graph", headers: { "Sec-Fetch-Site" => "same-origin" }
 
     assert_response :success
   end
@@ -30,7 +30,7 @@ class PrometheusProxyControllerTest < ActionDispatch::IntegrationTest
     sign_in
     on_subdomain :ams
 
-    get "/prometheus/graph"
+    get "/prometheus/graph", headers: { "Sec-Fetch-Site" => "same-origin" }
 
     assert_response :success
   end
@@ -66,13 +66,13 @@ class PrometheusProxyControllerTest < ActionDispatch::IntegrationTest
     sign_in
     on_subdomain :nyc
 
-    get "/prometheus/graph"
+    get "/prometheus/graph", headers: { "Sec-Fetch-Site" => "same-origin" }
 
     assert_response :not_found
   end
 
   test "proxy requires authentication" do
-    get "/prometheus/graph"
+    get "/prometheus/graph", headers: { "Sec-Fetch-Site" => "same-origin" }
 
     assert_response :redirect
     assert response.location.end_with?("/session/new")